+++ src/App/Scripting/ScriptingService.hpp
@@ -23,7 +23,7 @@ protected:
 
     static void OnInitializeScripts();
     static void OnInitializeGameInstance();
-    static void OnCreateInstance(Red::IScriptable*& aInstance, Red::CClass* aClass, uint32_t, bool);
+    static void OnCreateInstance(Red::IScriptable*& aInstance, Red::CClass* aClass, uint64_t, bool);
     static void OnValidateScripts(void* aValidator, Red::ScriptBundle* aBundle, void* aReport);
     static bool OnValidateProperty(void* aValidator, Red::ScriptProperty* aScriptProp, Red::CProperty* aNativeProp);
     static bool OnValidateScriptType(Red::CBaseRTTIType* aNativeType, Red::ScriptType* aScriptType);
